Speed up node install from ~60s to ~5s
If you download the tarball of the master branch, it’s throttled to like 50kb/sec, and downloading the tarball takes 20-60 seconds. If you download the tarball of the latest tagged release, it does not seem to be throttled at all, and reduces the amount of time the "install node" step takes on my machine down to about 5 seconds total.