Ensure that binfiles are files (not directories)
When there is a directory inside a Rails app's `bin/` (e.g.
`bin/stubs/`), it will print this after running
`rails g dockerfile`:

    file_clash  Dockerfile
    ...[rest of output removed]...

The `file_clash` is an odd error, but the root cause is trying to call
file-related methods on a directory (e.g. `IO.read(directory)`).

Now we filter the entries to only deal with actual files.